feat: CardRegister page storybook 작성 및 기존 App storybook 삭제
